rm_ta_time now supports trying to delete a non existing entry

+++ bota/bota/app_settings/views.py
@@ -174,6 +174,7 @@ def add_ta_time(request, course_id):
 @staff_member_required(login_url='/login/')
 def edit_ta_time(request, ta_id, course_id):
     tat = TATime.objects.get(id=ta_id)
+    print(tat.id)
     context = {'tat': tat,
                'date': str(tat.date),
                'start_time': str(tat.start_time),
@@ -196,7 +197,8 @@ def edit_ta_time(request, ta_id, course_id):
 
 @staff_member_required(login_url='/login/')
 def rm_ta_time(request, course_id, ta_id):
-    TATime.objects.get(id=ta_id).delete()
+    if (TATime.objects.filter(id=ta_id).exists()):
+        TATime.objects.get(id=ta_id).delete()
     return redirect('/settings/courses/'+course_id+'/edit')
 
 """
